bench: engine-agnostic restructure + third-party notice
Renames bench-duckdb-ab/ -> bench-external-ab/ and restructures the harness so external engines are adapters (harness/engines/fluree.py, engines/duckdb.py, engines/common.py); run_pair.py + check_pair.py are engine-agnostic cores that dispatch to the adapters. Adds the verbatim third-party NOTICE (nominative use; no DuckDB code/binaries/platform incorporated, external CLI invoked as a documented user install) at the top of README.md and PROTOCOL.md. Results language made factual (measurements, not competitive framing); where Fluree lags is stated plainly. Sanitization re-verified: only the <SNOWFLAKE_ACCOUNT> placeholder host remains.
