+++ bench-external-ab/harness/check_pair.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
-"""Correctness cross-check: run a pair on DuckDB (SQL) and Fluree (SPARQL) and compare
-results under the manifest's per-pair equivalence rule BEFORE any timing counts (PROTOCOL
-§6). A pair times only after it MATCHES here.
+"""Correctness cross-check: run a pair on the SQL engine and Fluree and compare results under
+the manifest's per-pair equivalence rule BEFORE any timing counts (PROTOCOL §6). A pair times
+only after it MATCHES here. Engine-specifics come from the adapters in engines/.
 
 Equivalence handling (PROTOCOL §7):
   - int/string cells      : exact multiset compare
@@ -9,8 +9,8 @@ Equivalence handling (PROTOCOL §7):
   - templated-IRI subject : reduced to its trailing integer key (so a Fluree IRI
                             http://data.fluree.dev/edw/store/3 compares to SQL store_key 3)
   - rows_only gate        : compare ROW COUNT only (unordered LIMIT)
-Gate per pair comes from pairs/manifest.json ("gate": exact|tolerance|rows_only) and
-"sum_is_double". Usage mirrors run_pair.py targets.
+  - columns aligned by NAME (SQL aliases == SPARQL vars), order-independent.
+Gate per pair comes from pairs/manifest.json ("gate": exact|tolerance|rows_only).
 
   python check_pair.py --pairs cq038_count_current_customers,cq014_channel_mix \\
     --duckdb-target duckdb-iceberg-minio-sf01 --fluree-target fluree-minio-sf01-main
@@ -20,6 +20,9 @@ import json
 import os
 import re
 import subprocess
+
+from engines import duckdb as duckdb_engine
+from engines import fluree as fluree_engine
 
 HERE = os.path.dirname(os.path.abspath(__file__))
 PAIRS = os.path.join(HERE, "pairs")
@@ -32,31 +35,6 @@ def load(p):
         return json.load(f)
 
 
-def bind_tables(sql, tgt):
-    mode = tgt["read_mode"]
-    def repl(m):
-        t = m.group(1).strip().lower()
-        if mode == "read_parquet":
-            return "read_parquet('%s/%s/data_0.parquet')" % (tgt["data_dir"], t)
-        return "%s.%s.%s" % (tgt["catalog_alias"], tgt["schema"], t.upper())
-    return re.sub(r"\{\{([^}]+)\}\}", repl, sql)
-
-
-def duckdb_preamble(tgt):
-    if tgt["read_mode"] != "iceberg_rest":
-        return ""
-    s = tgt["s3_secret"]
-    return ("INSTALL iceberg; LOAD iceberg; INSTALL httpfs; LOAD httpfs;\n"
-            "CREATE OR REPLACE SECRET s3sec (TYPE S3, KEY_ID '%s', SECRET '%s', ENDPOINT '%s', "
-            "URL_STYLE '%s', USE_SSL %s, REGION '%s');\n"
-            "ATTACH '%s' AS %s (TYPE ICEBERG, ENDPOINT '%s', AUTHORIZATION_TYPE %s, "
-            "ACCESS_DELEGATION_MODE %s);\n"
-            % (s["key_id"], s.get("secret_literal", ""), s["endpoint"], s.get("url_style", "path"),
-               str(s.get("use_ssl", False)).lower(), s.get("region", "us-east-1"),
-               tgt["warehouse"], tgt["catalog_alias"], tgt["endpoint"],
-               tgt.get("rest_auth", "none"), tgt.get("access_delegation", "none")))
-
-
 def norm_cell(v):
     """Normalize a raw string cell to a comparable python value."""
     if v is None or v == "":
@@ -65,18 +43,17 @@ def norm_cell(v):
     if m:
         return ("k", int(m.group(1)))
     try:
-        f = float(v)
-        return ("n", f)
+        return ("n", float(v))
     except ValueError:
         return ("s", v)
 
 
-def run_duckdb(pair, tgt):
-    """Returns (col_names, rows-as-dicts). Preamble output is sent to /dev/null so only the
-    query's header line + data land on stdout (header -> column names)."""
+def run_sql(pair, tgt):
+    """Returns (col_names, rows-as-dicts) from the SQL engine. Preamble output -> /dev/null so
+    only the query header line + data land on stdout (header -> column names)."""
     with open(os.path.join(PAIRS, pair + ".sql")) as f:
-        sql = bind_tables(f.read(), tgt)
-    script = (".mode csv\n.headers on\n.output /dev/null\n" + duckdb_preamble(tgt)
+        sql = duckdb_engine.bind_tables(f.read(), tgt)
+    script = (".mode csv\n.headers on\n.output /dev/null\n" + duckdb_engine.preamble(tgt)
               + ".output /dev/stdout\n" + sql.strip().rstrip(";") + ";\n")
     cmd = tgt.get("arch_prefix", []) + [tgt["bin"]]
     out = subprocess.run(cmd, input=script, capture_output=True, text=True, timeout=600).stdout
@@ -88,17 +65,11 @@ def run_duckdb(pair, tgt):
     return names, rows
 
 
-def run_fluree(pair, tgt):
-    """Returns (var_names, rows-as-dicts)."""
+def run_sparql(pair, tgt):
+    """Returns (var_names, rows-as-dicts) from Fluree."""
     with open(os.path.join(PAIRS, pair + ".rq")) as f:
         rq = f.read()
-    q = re.sub(r"(?is)\bWHERE\s*\{", "FROM <%s> WHERE {" % tgt["source"], rq, count=1)
-    env = os.environ.copy()
-    env.update(tgt.get("env", {}))
-    cmd = [tgt["fluree_bin"], "query", "--config", tgt["config"], "--connection",
-           "--sparql", "--format", "json", "-e", q]
-    out = subprocess.run(cmd, capture_output=True, text=True, timeout=600, env=env).stdout
-    doc = json.loads(out[out.index("{"):out.rindex("}") + 1])
+    doc = fluree_engine.query_json(tgt, rq)
     vars = doc["head"]["vars"]
     rows = [{v: (norm_cell(b[v]["value"]) if v in b else None) for v in vars}
             for b in doc["results"]["bindings"]]
@@ -126,7 +97,7 @@ def cmp_rows(dres, fres, gate):
     if gate == "rows_only":
         return len(dk) == len(fl), "rows d=%d f=%d" % (len(dk), len(fl))
     if set(dnames) != set(fnames):
-        return False, "COLUMN NAMES differ: duckdb=%s fluree=%s" % (sorted(dnames), sorted(fnames))
+        return False, "COLUMN NAMES differ: sql=%s fluree=%s" % (sorted(dnames), sorted(fnames))
     if len(dk) != len(fl):
         return False, "ROW COUNT d=%d f=%d" % (len(dk), len(fl))
     cols = sorted(set(dnames))  # align by NAME, order-independent
@@ -152,9 +123,7 @@ def main():
     for pair in [p.strip() for p in args.pairs.split(",") if p.strip()]:
         gate = man.get(pair, {}).get("gate", "exact")
         try:
-            dk = run_duckdb(pair, dt)
-            fl = run_fluree(pair, ft)
-            ok, msg = cmp_rows(dk, fl, gate)
+            ok, msg = cmp_rows(run_sql(pair, dt), run_sparql(pair, ft), gate)
         except Exception as e:
             ok, msg = False, "ERROR %s" % e
         allok = allok and ok
