+++ bench-external-ab/harness/pairs/cq008_revenue_by_region.sql
