+++ bench-external-ab/harness/pairs/cq027_web_events_by_type_device.sql
