+++ bench-external-ab/harness/pairs/cq038_count_current_customers.rq
