+++ bench-external-ab/harness/pairs/cq040_store_values.sql
