+++ bench-external-ab/harness/pairs/crt_highcard_groupby.sql
