+++ bench-external-ab/harness/pairs/crt_minmax.sql
