+++ bench-external-ab/harness/pairs/p1_count_fact.sql
