+++ bench-external-ab/harness/pairs/p2_category_rollup.rq
