+++ bench-external-ab/harness/pairs/p3_open_tickets_by_segment.sql
